requirements.txt
torch
transformers
datasets
tokenizers
accelerate
fastapi
uvicorn
python-multipart
pydantic
pyyaml
sentencepiece
tiktoken
